app/requirements.txt
# Requirements for Databricks Apps deployment
# These are the dependencies needed to run the Streamlit app

streamlit>=1.40.0
databricks-sdk>=0.38.0
pydantic>=2.0
python-dotenv
mlflow>=3.6.0
openai>=1.0
requests>=2.31.0

